Rename ctz and popcnt source file, comment build constraints
