+++ popcnt_compat.go
