+++ popcnt_slices_test.go
